storybook: move sources
